requirements.txt
loguru
slackclient
python-dotenv